fix: ProbeForLandlock falsely reports landlock on macOS
On non-Linux platforms, ProbeLandlockSelfDepth() returned (0, nil)
which ProbeForLandlock() misinterpreted as 'landlocked' (depth 0 != MAX_DEPTH 16).

Fix by splitting ProbeForLandlock into a Linux-only implementation
(landlock_linux.go) and a !linux stub returning (false, nil) in
landlock.go, mirroring the existing pattern for ProbeLandlockSelfDepth.

Also move /proc-dependent tests (getHostMounts, getRunningProcess,
getRunningParentProcess, getRunningProcesses) into *_linux_test.go
files so they are skipped on macOS, and update go.sum via proxy.golang.org
to resolve gitleaks v8.30.1 checksum mismatch.

Add GOPROXY=https://proxy.golang.org,direct to flake.nix devshell env
to prevent the checksum mismatch from recurring.
