+++ pkg/tasks/baseline/environment_test.go
@@ -142,8 +142,4 @@ func Test_getContainerRuntime(t *testing.T) {
 	}
 }
 
-func Test_getHostMounts(t *testing.T) {
-	mounts, err := GetHostMounts()
-	t.Logf("Mounts found: %v", mounts)
-	assert.NoError(t, err)
-}
+
