+++ pkg/tasks/baseline/landlock.go
@@ -6,3 +6,7 @@ package tasks
 func ProbeLandlockSelfDepth() (int, error) {
 	return 0, nil
 }
+
+func ProbeForLandlock() (bool, error) {
+	return false, nil
+}
