+++ pkg/tasks/baseline/landlock_linux.go
@@ -1,12 +1,47 @@
+//go:build linux
+// +build linux
+
 package tasks
 
 import (
+	"encoding/json"
 	"errors"
 	"fmt"
+	"os"
+	"os/exec"
 	"unsafe"
 
 	"golang.org/x/sys/unix"
 )
+
+func ProbeForLandlock() (bool, error) {
+	self, err := os.Executable()
+	if err != nil {
+		return false, fmt.Errorf("failed to get self executable location: %w", err)
+	}
+
+	cmd := exec.Command(self, "probe")
+	out, err := cmd.Output()
+	if err != nil {
+		return false, err
+	}
+
+	var pd ProbeSubCmdData
+	err = json.Unmarshal(out, &pd)
+	if err != nil {
+		return false, err
+	}
+
+	if pd.LockdownDepth > LANDLOCK_MAX_DEPTH {
+		return false, fmt.Errorf("lockdown depth isn't expected to ever go beyond %d", LANDLOCK_MAX_DEPTH)
+	}
+
+	if pd.LockdownDepth == LANDLOCK_MAX_DEPTH {
+		return false, nil
+	}
+
+	return true, nil
+}
 
 func ProbeLandlockSelfDepth() (int, error) {
 	// Required before landlock_restrict_self
